Bakalar-Mark-Vlkolensky-kody/Bakalar/MicroserviceEDA/DatabaseMicroserviceEDA/JeevesDatabase/src/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiokafka==0.7.2
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-WTF==1.0.0
frozenlist==1.3.0
greenlet==1.1.2
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.13.2
hyperframe==6.0.1
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.1
kafka-python==2.0.2
MarkupSafe==2.1.1
multidict==6.0.2
priority==2.0.0
quart==0.17.0
requests==2.27.1
SQLAlchemy==1.4.32
toml==0.10.2
typing_extensions==4.1.1
urllib3==1.26.9
Werkzeug==2.0.3
wsproto==1.1.0
WTForms==3.0.1
yarl==1.7.2
